Hex control improvements

+++ TotalPE/DataDirectoriesView.cpp
@@ -49,7 +49,7 @@ void CDataDirectoriesView::OnStateChanged(HWND, int from, int to, DWORD oldState
 		int selected = m_List.GetSelectedCount();
 		if (selected == 1) {
 			auto const& dir = m_Directories[m_List.GetNextItem(-1, LVNI_SELECTED)];
-			auto offset = m_PE.GetOffsetFromRVA(dir.DataDir.VirtualAddress);
+			auto offset = (DWORD)m_PE.GetOffsetFromRVA(dir.DataDir.VirtualAddress);
 			m_HexView.SetData(m_PE, offset, dir.DataDir.Size);
 		}
 		else {
